Added relationship

+++ app/routers/film.py
@@ -12,22 +12,23 @@ router = APIRouter(
 )
 
 @router.get("/", response_model=List[schemas.FilmSchemaResponse])
-def read_films(db: Session = Depends(get_db)):
-    statement = select(models.Film)
+def read_films(db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user),
+               limit: int = 10):
+    statement = select(models.Film).where(models.Film.user_id == current_user.id).limit(limit)
     result = db.execute(statement).scalars().all()
     return result
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.FilmSchemaResponse)
 def create_films(film: schemas.CreateFilm, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
-    new_film = models.Film(**film.model_dump())
+    new_film = models.Film(user_id=current_user.id, **film.model_dump())
     db.add(new_film)
     db.commit()
     db.refresh(new_film)
     return new_film
 
 @router.get("/{id}", response_model=schemas.FilmSchemaResponse)
-def read_film_by_id(id: int, db: Session = Depends(get_db)):
-    statement = select(models.Film).where(models.Film.id == id)
+def read_film_by_id(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
+    statement = select(models.Film).where(models.Film.id == id and models.Film.user_id == current_user.id)
     result = db.execute(statement).scalars().one_or_none()
     if not result:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Film with id: {id} was not found")
@@ -35,22 +36,29 @@ def read_film_by_id(id: int, db: Session = Depends(get_db)):
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
 def delete_film(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
-    # Find the film by id
     statement = select(models.Film).where(models.Film.id == id)
     film_to_delete = db.execute(statement).scalars().one_or_none()
     if film_to_delete is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Film with id: {id} does not exist")
+    if film_to_delete.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorised to perform requested action")
     db.delete(film_to_delete)
     db.commit()
 
 @router.put("/{id}", response_model=schemas.FilmSchemaResponse)
 def update_film(id: int, film: schemas.UpdateFilm, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     updated_film_data = film.model_dump(exclude_unset=True)
-    statement = update(models.Film).where(models.Film.id == id).values(**updated_film_data)
-    result = db.execute(statement)
-    if result.rowcount == 0:
+    existing_film = db.execute(
+        select(models.Film).where(models.Film.id == id)
+    ).scalars().one_or_none()
+    if not existing_film:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Film with id: {id} does not exist")
+    if existing_film.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorised to perform requested action")
+    statement = update(models.Film).where(models.Film.id == id).values(**updated_film_data)
+    db.execute(statement)
     db.commit()
-    statement = select(models.Film).where(models.Film.id == id)
-    updated_film = db.execute(statement).scalars().first()
+    updated_film = db.execute(
+        select(models.Film).where(models.Film.id == id)
+    ).scalars().first()
     return updated_film
